+++ assets/js/app.js
@@ -14,6 +14,7 @@ import toast from './ui/toast.js'
 import sortable from './ui/sortable.js';
 import collapseSubSidebar from './ui/collapsible_sub_sidebar.js'
 import fileUpload from './ui/file_upload.js'
+import formAutoFocus from './ui/form_auto_focus.js'
 
 // configure htmx
 htmx.config.defaultFocusScroll = true
@@ -38,4 +39,5 @@ document.addEventListener('DOMContentLoaded', function () {
     sortable()
     collapseSubSidebar()
     fileUpload()
+    formAutoFocus()
 });